# redis-trib: use pipeline to speed up moving slot