Closes #2710

* "keysinslot" now retrieves more keys (10 -> 10000) for every iteration
* Use pipeline to run "migrate" commands much faster then current

Test environment is below,
- Tested on my MacBook Pro Retina. (15 inch, Late 2013)
  - OSX 10.10.3
  - i7 2.3GHz, 16GB DDR3 1600 MHz, SSD 256GB
- total 6 processes of redis-server (cluster mode)
  - 3 masters, 3 slaves
- 1000001 keys in one slot (slot number 13784)
  - “{hashtag}0” to “{hashtag}1000000”
- check elapsed seconds for only moving slot 13784

Test result is here,
a) verbose = true
- origin (keysinslot 10, not using pipeline) : elapsed: 142.083674 secs
- keysinslot 1000, using pipeline : elapsed: 59.248071 secs
- keysinslot 10000, using pipeline : elapsed: 57.582782 secs

b) verbose = false
- origin (keysinslot 10, not using pipeline) : elapsed: 115.745819 secs
- keysinslot 1000, using pipeline : elapsed: 59.676552 secs
- keysinslot 10000, using pipeline : elapsed: 57.819602 secs

Increasing fetch size of keysinslot and using pipeline seems to be 2 ~ 2.5 times faster than current.
I couldn't test it with multiple machines since I don't have three or more physical machines.

Btw, just modifying verbose option to 'false' also reduces more than 25 secs.