@Calabor-Hoc 2015-08-07
That's great! But it's a little disappointing that pipeline can only shorten the time for half. I thought the pipeline have much more potential...