@HeartSaVioR 2015-08-07
@Calabor-Hoc 
I don't know about how redis-rb pipeline works, but in benchmark, Jedis pipelining is more than 15x faster than normal.
https://github.com/xetorthio/jedis/pull/713#issuecomment-66904746