@antirez 2015-12-11
Multi keys `MIGRATE` is now merged into `unstable`. I'm implementing it into redis-trib (trivial) to check the gain (harder, benchmarking is always hard).